+++ src/main/java/io/github/hee9841/excel/annotation/Excel.java
@@ -41,7 +41,7 @@ import java.lang.annotation.Target;
  * 
  * // Usage example:
  * List{@literal <}UserData{@literal >} users = getUserData();
- * DefaultExcelExporter{@literal <}UserData{@literal >} exporter = DefaultExcelExporter.builder(UserData.class, users)
+ * SXSSFExporter{@literal <}UserData{@literal >} exporter = SXSSFExporter.builder(UserData.class, users)
  *     .sheetName("Users")
  *     .build();
  * exporter.write(new FileOutputStream("users.xlsx"));
